\name{edaplots}
\alias{edaplots}
\title{edaplots}
\usage{
This is very useful function to create all the required plots for eda

}
\description{
This function takes input as a dataframe and creates barplots, piecharts, histograms, boxplots for univariate analysis and it also makes bivariate boxplots and scatterplots for bivariate analysis. Plots are exported to respective folders in working directory as .png files
}
\examples{edaplot(data,'dataname')

}
